added toggleMode button in the header and removed personalize settings from settings
