+++ client/index.html
@@ -6,7 +6,7 @@
     <link rel="shortcut icon" href="/media/images/woman-sigh.png" type="image/x-icon">
     <title>DIGITAL PROGRESS CARD</title>
   </head>
-  <body class="light">
+  <body>
     <div id="root"></div>
     <script type="module" src="/src/main.jsx"></script>
   </body>
