+++ client/src/components/Settings/Settings.jsx
@@ -8,26 +8,26 @@ const Settings = () => {
       <div className="flex flex-col gap-4">
         <NavLink
           to="/settings/account"
-          className="flex transform items-center rounded-lg px-3 py-2 text-gray-600 transition-colors duration-300 hover:bg-gray-100 hover:text-gray-700"
+          className="flex transform items-center rounded-lg px-3 py-2 text-gray-600 transition-colors duration-300  hover:text-black"
 
         >
           <h2 className="text-xl font-semibold mb-2">Account Settings</h2>
           
         </NavLink>
-        <NavLink
+        {/* <NavLink
           to="/settings/personalize"
           className="flex transform items-center rounded-lg px-3 py-2 text-gray-600 transition-colors duration-300 hover:bg-gray-100 hover:text-gray-700"
         >
           <h2 className="text-xl font-semibold mb-2">Personalize Settings</h2>
           
-        </NavLink>
-        <NavLink
+        </NavLink> */}
+        {/* <NavLink
           to="/settings"
           className="flex transform items-center rounded-lg px-3 py-2 text-gray-600 transition-colors duration-300 hover:bg-gray-100 hover:text-gray-700"
         >
           <h2 className="text-xl font-semibold mb-2">Other Settings</h2>
           
-        </NavLink>
+        </NavLink> */}
       </div>
     </div>
   );
