+++ client/src/styles/index.css
@@ -14,8 +14,8 @@
 }
 
 .dark{
-    --primary-color:#333;
-    --accent-color:#111b20;
+    --primary-color:#353535;
+    --accent-color:#000;
     --opt-color:#1b4332;
     --hover-color:#5f5f5f;
     --secondry-color:#1b4332;
@@ -25,7 +25,7 @@
     --third-font: 'Ubuntu', sans-serif;
 }
 
-#blue{
+/* #blue{
     --primary-color:blue;
     --accent-color:#111b20;
     --opt-color:#1b4332;
@@ -79,7 +79,7 @@
     --text-color: #000000 --primary-font: 'Rubik', sans-serif;
     --secondary-font: 'Roboto', sans-serif;
     --third-font: 'Ubuntu', sans-serif;
-}
+} */
 
 .small{
     font-size: 11px;
