+++ client/src/styles/toggleMode.css
@@ -9,13 +9,13 @@
     border-color: var(--hover-color); /* bg-gray-300 */
   }
   
-  .toggle-checkbox:checked + .toggle-label {
-    background-color: black; /* bg-gray-800 */
+  .toggle-checkbox:checked + .toggle-label{
+    background-color: rgb(255, 255, 255); /* bg-gray-800 */
   }
   
   .toggle-checkbox:checked + .toggle-label::before {
     transform: translateX(1.25rem); /* Adjust according to the size of your switch */
-    background-color: white; /* bg-gray-800 */
+    background-color: black; /* bg-gray-800 */
   }
   
   .toggle-label {
@@ -24,9 +24,9 @@
     cursor: pointer;
     border: 2px solid transparent;
     border-radius: 9999px;
-    height: 1.25rem;
+    height: 1.25em;
     width: 2.5rem;
-    background-color: rgba(243, 239, 24, 0.829); /* bg-gray-300 */
+    background-color: yellow; /* bg-gray-300 */
     transition: background-color 0.2s ease-in-out;
   }
   
@@ -36,9 +36,10 @@
     height: 1rem;
     width: 1rem;
     border-radius: 9999px;
-    background: var(--light-text);
-    box-shadow: 0 0 2px 0 rgba(0, 0, 0, 0.2);
+    /* background: var(--light-text); */
+    box-shadow: 0 0 2px 0 rgba(0, 0, 0, 0.295);
     transition: transform 0.2s ease-in-out;
+    background-color: rgb(255, 255, 255);
   }
   .toggle-label::after{
     background: var(--light-text);
